Behaviour_Simulation/requirements.txt
numpy 
pandas 
scikit-learn
tensorflow
emoji
requests
opencv-python
transformers 
Pillow 
gensim
nltk 
gdown